# openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: TZFService API
    description: Simple server convert longitude&latitude to timezone name
    version: 0.0.1
paths:
    /api/v1/tz:
        get:
            tags:
                - TZFService
            operationId: TZFService_GetTimezone
            parameters:
                - name: longitude
                  in: query
                  description: '@gotags: query:"lng" vd:"$>=-180 && $<=180"'
                  schema:
                    type: number
                    format: double
                - name: latitude
                  in: query
                  description: '@gotags: query:"lat" vd:"$>=-90 && $<=90"'
                  schema:
                    type: number
                    format: double
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetTimezoneResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/tzs:
        get:
            tags:
                - TZFService
            operationId: TZFService_GetTimezones
            parameters:
                - name: longitude
                  in: query
                  description: '@gotags: query:"lng" vd:"$>=-180 && $<=180"'
                  schema:
                    type: number
                    format: double
                - name: latitude
                  in: query
                  description: '@gotags: query:"lat" vd:"$>=-90 && $<=90"'
                  schema:
                    type: number
                    format: double
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetTimezonesResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/tzs/all:
        get:
            tags:
                - TZFService
            operationId: TZFService_GetAllTimezones
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetAllTimezonesResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        GetAllTimezonesResponse:
            type: object
            properties:
                timezones:
                    type: array
                    items:
                        $ref: '#/components/schemas/GetTimezoneResponse'
        GetTimezoneResponse:
            type: object
            properties:
                timezone:
                    type: string
                abbreviation:
                    type: string
                offset:
                    pattern: ^-?(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,9})?s$
                    type: string
        GetTimezonesResponse:
            type: object
            properties:
                timezones:
                    type: array
                    items:
                        $ref: '#/components/schemas/GetTimezoneResponse'
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
tags:
    - name: TZFService